# swagger.json
{"openapi":"3.1.0","info":{"title":"Titanic API","description":"API for Titanic passengers data","version":"1.1.0"},"paths":{"/":{"get":{"summary":"Read Root","description":"This function handles GET requests to the root URL (\"/\").\nIt returns a welcome message.","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/fare_histogram":{"get":{"summary":"Fare Histogram","description":"This function handles GET requests to the \"/fare_histogram\" URL.\nIt calculates and returns a histogram data of passenger fares.\nRendering the response on server side is not optimal due to traffic overhead,\ntherefore the rendering can be done on client side and the response body will contain only the numerical histogram data.","operationId":"fare_histogram_fare_histogram_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/passenger/{passenger_id}":{"get":{"summary":"Passenger","description":"This function handles GET requests to the \"/passenger/{passenger_id}\" URL.\nIt returns the data for the passenger with the given ID.\nIf the \"attributes\" query parameter is provided, only the specified attributes are returned.","operationId":"passenger_passenger__passenger_id__get","parameters":[{"name":"passenger_id","in":"path","required":true,"schema":{"type":"integer","title":"Passenger Id"}},{"name":"attributes","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Attributes"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/passengers":{"get":{"summary":"Passengers","description":"This function handles GET requests to the \"/passengers\" URL.\nIt returns the data for all passengers.","operationId":"passengers_passengers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}